+++ lib/data_store/table.rb
@@ -1,10 +1,10 @@
 module DataStore
 
-  class Stack
+  class Table
 
     attr_reader :identifier
 
-    # Initialize the stack by passsing an identifier
+    # Initialize the table by passsing an identifier
     def initialize(identifier)
       @identifier = identifier
     end
@@ -14,13 +14,13 @@ module DataStore
       @parent ||= DataStore::Base.find(identifier: identifier)
     end
 
-    # Return a Stack object enriched with Sequel::Model behaviour
+    # Return a table object enriched with Sequel::Model behaviour
     def model
       @model ||= Class.new(Sequel::Model(dataset))
     end
 
-    # Push a new datapoint on the stack
-    def push(value)
+    # Add a new datapoint on the table
+    def add(value)
       dataset << {value: value, created: Time.now.utc.to_f}
       calculate_average_values
     end
@@ -30,12 +30,12 @@ module DataStore
       model.order(:created).last
     end
 
-    # Return the total number of datapoints in the stack
+    # Return the total number of datapoints in the table
     def count
       dataset.count
     end
 
-    # Create the database tables which the stack usesd for storing the datapoints
+    # Create the database tables which the table usesd for storing the datapoints
     def create!
       migrate(:up)
     end
@@ -48,7 +48,7 @@ module DataStore
 
     # Return the corresponding dataset with the datapoitns
     def dataset
-      database[stack_name]
+      database[table_name]
     end
 
     private
@@ -58,20 +58,20 @@ module DataStore
       calculator.perform
     end
 
-    def stack_table_names
-      names  = [stack_name]
+    def table_names
+      names  = [table_name]
       factor = 1
       parent.compression_schema.each do |compression|
         factor = (factor * compression)
-        names << stack_name.to_s + '_' + factor.to_s
+        names << table_name.to_s + '_' + factor.to_s
       end
       names
     end
 
     def migrate(direction = :up)
-      stack_table_names.each do |name|
+      table_names.each do |name|
         begin
-          DataStore.create_stack(name).apply(database, direction)
+          DataStore.create_table(name).apply(database, direction)
         rescue Sequel::DatabaseError
         end
       end
@@ -81,7 +81,7 @@ module DataStore
       @database ||= DataStore::Base.db
     end
 
-    def stack_name
+    def table_name
       (prefix + identifier.to_s).to_sym
     end