+++ CHANGELOG.md
@@ -3,7 +3,7 @@
   * Readme driven development
   * Configuration of DataStore with Connector class for database connection and dataset definition
   * Introduction of DataStore::Base enriched behaviour with the use of Sequel::Model
-  * introduction of the DataStore::Stack to push datapoints
+  * introduction of the DataStore::Table to add datapoints
 
 ##0.0.1